# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Dashboard API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type SignalAssetClass =
  (typeof SignalAssetClass)[keyof typeof SignalAssetClass];

export const SignalAssetClass = {
  stocks: "stocks",
  indices: "indices",
  forex: "forex",
} as const;

export type SignalSignalType =
  (typeof SignalSignalType)[keyof typeof SignalSignalType];

export const SignalSignalType = {
  buy: "buy",
  sell: "sell",
  hold: "hold",
} as const;

export interface Signal {
  id: number;
  symbol: string;
  name: string;
  assetClass: SignalAssetClass;
  signalType: SignalSignalType;
  confidence: number;
  currentPrice: number;
  targetPrice: number;
  stopLoss: number;
  riskReward: number;
  timeframe: string;
  reasoning: string;
  generatedAt: string;
}

export type NewSignalAssetClass =
  (typeof NewSignalAssetClass)[keyof typeof NewSignalAssetClass];

export const NewSignalAssetClass = {
  stocks: "stocks",
  indices: "indices",
  forex: "forex",
} as const;

export type NewSignalSignalType =
  (typeof NewSignalSignalType)[keyof typeof NewSignalSignalType];

export const NewSignalSignalType = {
  buy: "buy",
  sell: "sell",
  hold: "hold",
} as const;

export interface NewSignal {
  symbol: string;
  name: string;
  assetClass: NewSignalAssetClass;
  signalType: NewSignalSignalType;
  confidence: number;
  currentPrice: number;
  targetPrice: number;
  stopLoss: number;
  riskReward: number;
  timeframe: string;
  reasoning: string;
}

export type SignalSummaryByAssetClass = {
  stocks: number;
  indices: number;
  forex: number;
};

export interface SignalSummary {
  total: number;
  buy: number;
  sell: number;
  hold: number;
  byAssetClass: SignalSummaryByAssetClass;
  avgConfidence: number;
}

export interface Portfolio {
  totalValue: number;
  cashBalance: number;
  investedValue: number;
  totalPnl: number;
  totalPnlPercent: number;
  dayPnl: number;
  dayPnlPercent: number;
  openPositions: number;
  winRate: number;
  totalTrades: number;
}

export interface PortfolioDataPoint {
  date: string;
  value: number;
}

export type PositionAssetClass =
  (typeof PositionAssetClass)[keyof typeof PositionAssetClass];

export const PositionAssetClass = {
  stocks: "stocks",
  indices: "indices",
  forex: "forex",
} as const;

export type PositionDirection =
  (typeof PositionDirection)[keyof typeof PositionDirection];

export const PositionDirection = {
  long: "long",
  short: "short",
} as const;

export interface Position {
  id: number;
  symbol: string;
  name: string;
  assetClass: PositionAssetClass;
  direction: PositionDirection;
  quantity: number;
  entryPrice: number;
  currentPrice: number;
  targetPrice: number;
  stopLoss: number;
  pnl: number;
  pnlPercent: number;
  openedAt: string;
}

export type NewPositionAssetClass =
  (typeof NewPositionAssetClass)[keyof typeof NewPositionAssetClass];

export const NewPositionAssetClass = {
  stocks: "stocks",
  indices: "indices",
  forex: "forex",
} as const;

export type NewPositionDirection =
  (typeof NewPositionDirection)[keyof typeof NewPositionDirection];

export const NewPositionDirection = {
  long: "long",
  short: "short",
} as const;

export interface NewPosition {
  symbol: string;
  name: string;
  assetClass: NewPositionAssetClass;
  direction: NewPositionDirection;
  quantity: number;
  entryPrice: number;
  targetPrice: number;
  stopLoss: number;
}

export type AssetAssetClass =
  (typeof AssetAssetClass)[keyof typeof AssetAssetClass];

export const AssetAssetClass = {
  stocks: "stocks",
  indices: "indices",
  forex: "forex",
} as const;

export interface Asset {
  id: number;
  symbol: string;
  name: string;
  assetClass: AssetAssetClass;
  currentPrice: number;
  change: number;
  changePercent: number;
  volume: string;
  marketCap: string;
}

export interface ApiKey {
  id: number;
  name: string;
  /** First 8 chars of key for identification */
  prefix: string;
  active: boolean;
  lastUsedAt?: string | null;
  createdAt: string;
}

export interface ApiKeyCreated {
  id: number;
  name: string;
  prefix: string;
  /** Full API key — only shown once at creation */
  key: string;
  active: boolean;
  createdAt: string;
}

export interface CreateApiKeyBody {
  name: string;
}

export interface IngestionRules {
  id: number;
  /** Minimum confidence % to accept signal (0-100) */
  minConfidence: number;
  allowStocks: boolean;
  allowIndices: boolean;
  allowForex: boolean;
  requireReasoning: boolean;
  requireStopLoss: boolean;
  requireTargetPrice: boolean;
  updatedAt: string;
}

export interface UpdateIngestionRulesBody {
  minConfidence: number;
  allowStocks: boolean;
  allowIndices: boolean;
  allowForex: boolean;
  requireReasoning: boolean;
  requireStopLoss: boolean;
  requireTargetPrice: boolean;
}

export type IngestionLogEntrySource =
  (typeof IngestionLogEntrySource)[keyof typeof IngestionLogEntrySource];

export const IngestionLogEntrySource = {
  agent: "agent",
  tradingview: "tradingview",
  manual: "manual",
} as const;

export interface IngestionLogEntry {
  id: number;
  source: IngestionLogEntrySource;
  symbol: string;
  signalType: string;
  accepted: boolean;
  rejectionReason?: string | null;
  keyName?: string | null;
  createdAt: string;
}

export interface TradingViewPayload {
  ticker: string;
  /** buy, sell, or hold */
  action: string;
  price: number;
  confidence?: number;
  target?: number;
  stop?: number;
  message?: string;
  timeframe?: string;
}

export type GetSignalsParams = {
  asset_class?: GetSignalsAssetClass;
  signal_type?: GetSignalsSignalType;
};

export type GetSignalsAssetClass =
  (typeof GetSignalsAssetClass)[keyof typeof GetSignalsAssetClass];

export const GetSignalsAssetClass = {
  stocks: "stocks",
  indices: "indices",
  forex: "forex",
  all: "all",
} as const;

export type GetSignalsSignalType =
  (typeof GetSignalsSignalType)[keyof typeof GetSignalsSignalType];

export const GetSignalsSignalType = {
  buy: "buy",
  sell: "sell",
  hold: "hold",
  all: "all",
} as const;

export type GetIngestionLogParams = {
  status?: GetIngestionLogStatus;
  limit?: number;
};

export type GetIngestionLogStatus =
  (typeof GetIngestionLogStatus)[keyof typeof GetIngestionLogStatus];

export const GetIngestionLogStatus = {
  all: "all",
  accepted: "accepted",
  rejected: "rejected",
} as const;