# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Trading Signal Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: signals
    description: Trading signals
  - name: portfolio
    description: Portfolio management
  - name: positions
    description: Open positions
  - name: assets
    description: Watchlist assets
  - name: keys
    description: API key management
  - name: ingestion
    description: Signal ingestion rules and log
  - name: webhooks
    description: Webhook endpoints

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /signals:
    get:
      operationId: getSignals
      tags: [signals]
      summary: Get all trading signals
      parameters:
        - name: asset_class
          in: query
          schema:
            type: string
            enum: [stocks, indices, forex, all]
          required: false
        - name: signal_type
          in: query
          schema:
            type: string
            enum: [buy, sell, hold, all]
          required: false
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
    post:
      operationId: createSignal
      tags: [signals]
      summary: Create a new trading signal (requires X-API-Key header)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSignal"
      responses:
        "201":
          description: Signal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "401":
          description: Missing or invalid API key
        "422":
          description: Validation rules rejected the signal

  /signals/{id}:
    get:
      operationId: getSignal
      tags: [signals]
      summary: Get a single signal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
    delete:
      operationId: deleteSignal
      tags: [signals]
      summary: Delete a signal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /signals/summary:
    get:
      operationId: getSignalSummary
      tags: [signals]
      summary: Get signal summary counts by type and asset class
      responses:
        "200":
          description: Signal summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalSummary"

  /portfolio:
    get:
      operationId: getPortfolio
      tags: [portfolio]
      summary: Get portfolio overview
      responses:
        "200":
          description: Portfolio overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"

  /portfolio/history:
    get:
      operationId: getPortfolioHistory
      tags: [portfolio]
      summary: Get portfolio value history for chart
      responses:
        "200":
          description: Portfolio history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioDataPoint"

  /positions:
    get:
      operationId: getPositions
      tags: [positions]
      summary: Get all open positions
      responses:
        "200":
          description: List of positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
    post:
      operationId: createPosition
      tags: [positions]
      summary: Open a new position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewPosition"
      responses:
        "201":
          description: Position opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"

  /positions/{id}:
    delete:
      operationId: closePosition
      tags: [positions]
      summary: Close a position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Closed

  /assets:
    get:
      operationId: getAssets
      tags: [assets]
      summary: Get watchlist assets with latest price info
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"

  /keys:
    get:
      operationId: getApiKeys
      tags: [keys]
      summary: List all API keys (values are masked)
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [keys]
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyBody"
      responses:
        "201":
          description: API key created — full key only shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"

  /keys/{id}:
    delete:
      operationId: deleteApiKey
      tags: [keys]
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Key revoked

  /ingestion/rules:
    get:
      operationId: getIngestionRules
      tags: [ingestion]
      summary: Get current signal validation rules
      responses:
        "200":
          description: Current rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionRules"
    put:
      operationId: updateIngestionRules
      tags: [ingestion]
      summary: Update signal validation rules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIngestionRulesBody"
      responses:
        "200":
          description: Updated rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionRules"

  /ingestion/log:
    get:
      operationId: getIngestionLog
      tags: [ingestion]
      summary: Get signal ingestion audit log
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, accepted, rejected]
          required: false
        - name: limit
          in: query
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Ingestion log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IngestionLogEntry"

  /webhooks/tradingview:
    post:
      operationId: tradingViewWebhook
      tags: [webhooks]
      summary: Receive a TradingView alert and convert to signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradingViewPayload"
      responses:
        "201":
          description: Signal created from webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "401":
          description: Missing or invalid API key
        "422":
          description: Validation rules rejected the signal

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Signal:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        assetClass:
          type: string
          enum: [stocks, indices, forex]
        signalType:
          type: string
          enum: [buy, sell, hold]
        confidence:
          type: number
        currentPrice:
          type: number
        targetPrice:
          type: number
        stopLoss:
          type: number
        riskReward:
          type: number
        timeframe:
          type: string
        reasoning:
          type: string
        generatedAt:
          type: string
          format: date-time
      required:
        - id
        - symbol
        - name
        - assetClass
        - signalType
        - confidence
        - currentPrice
        - targetPrice
        - stopLoss
        - riskReward
        - timeframe
        - reasoning
        - generatedAt

    NewSignal:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        assetClass:
          type: string
          enum: [stocks, indices, forex]
        signalType:
          type: string
          enum: [buy, sell, hold]
        confidence:
          type: number
        currentPrice:
          type: number
        targetPrice:
          type: number
        stopLoss:
          type: number
        riskReward:
          type: number
        timeframe:
          type: string
        reasoning:
          type: string
      required:
        - symbol
        - name
        - assetClass
        - signalType
        - confidence
        - currentPrice
        - targetPrice
        - stopLoss
        - riskReward
        - timeframe
        - reasoning

    SignalSummary:
      type: object
      properties:
        total:
          type: integer
        buy:
          type: integer
        sell:
          type: integer
        hold:
          type: integer
        byAssetClass:
          type: object
          properties:
            stocks:
              type: integer
            indices:
              type: integer
            forex:
              type: integer
          required:
            - stocks
            - indices
            - forex
        avgConfidence:
          type: number
      required:
        - total
        - buy
        - sell
        - hold
        - byAssetClass
        - avgConfidence

    Portfolio:
      type: object
      properties:
        totalValue:
          type: number
        cashBalance:
          type: number
        investedValue:
          type: number
        totalPnl:
          type: number
        totalPnlPercent:
          type: number
        dayPnl:
          type: number
        dayPnlPercent:
          type: number
        openPositions:
          type: integer
        winRate:
          type: number
        totalTrades:
          type: integer
      required:
        - totalValue
        - cashBalance
        - investedValue
        - totalPnl
        - totalPnlPercent
        - dayPnl
        - dayPnlPercent
        - openPositions
        - winRate
        - totalTrades

    PortfolioDataPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
      required:
        - date
        - value

    Position:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        assetClass:
          type: string
          enum: [stocks, indices, forex]
        direction:
          type: string
          enum: [long, short]
        quantity:
          type: number
        entryPrice:
          type: number
        currentPrice:
          type: number
        targetPrice:
          type: number
        stopLoss:
          type: number
        pnl:
          type: number
        pnlPercent:
          type: number
        openedAt:
          type: string
          format: date-time
      required:
        - id
        - symbol
        - name
        - assetClass
        - direction
        - quantity
        - entryPrice
        - currentPrice
        - targetPrice
        - stopLoss
        - pnl
        - pnlPercent
        - openedAt

    NewPosition:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        assetClass:
          type: string
          enum: [stocks, indices, forex]
        direction:
          type: string
          enum: [long, short]
        quantity:
          type: number
        entryPrice:
          type: number
        targetPrice:
          type: number
        stopLoss:
          type: number
      required:
        - symbol
        - name
        - assetClass
        - direction
        - quantity
        - entryPrice
        - targetPrice
        - stopLoss

    Asset:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        assetClass:
          type: string
          enum: [stocks, indices, forex]
        currentPrice:
          type: number
        change:
          type: number
        changePercent:
          type: number
        volume:
          type: string
        marketCap:
          type: string
      required:
        - id
        - symbol
        - name
        - assetClass
        - currentPrice
        - change
        - changePercent
        - volume
        - marketCap

    ApiKey:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        prefix:
          type: string
          description: First 8 chars of key for identification
        active:
          type: boolean
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - prefix
        - active
        - createdAt

    ApiKeyCreated:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        prefix:
          type: string
        key:
          type: string
          description: Full API key — only shown once at creation
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - prefix
        - key
        - active
        - createdAt

    CreateApiKeyBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    IngestionRules:
      type: object
      properties:
        id:
          type: integer
        minConfidence:
          type: number
          description: Minimum confidence % to accept signal (0-100)
        allowStocks:
          type: boolean
        allowIndices:
          type: boolean
        allowForex:
          type: boolean
        requireReasoning:
          type: boolean
        requireStopLoss:
          type: boolean
        requireTargetPrice:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - minConfidence
        - allowStocks
        - allowIndices
        - allowForex
        - requireReasoning
        - requireStopLoss
        - requireTargetPrice
        - updatedAt

    UpdateIngestionRulesBody:
      type: object
      properties:
        minConfidence:
          type: number
        allowStocks:
          type: boolean
        allowIndices:
          type: boolean
        allowForex:
          type: boolean
        requireReasoning:
          type: boolean
        requireStopLoss:
          type: boolean
        requireTargetPrice:
          type: boolean
      required:
        - minConfidence
        - allowStocks
        - allowIndices
        - allowForex
        - requireReasoning
        - requireStopLoss
        - requireTargetPrice

    IngestionLogEntry:
      type: object
      properties:
        id:
          type: integer
        source:
          type: string
          enum: [agent, tradingview, manual]
        symbol:
          type: string
        signalType:
          type: string
        accepted:
          type: boolean
        rejectionReason:
          type: string
          nullable: true
        keyName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - source
        - symbol
        - signalType
        - accepted
        - createdAt

    TradingViewPayload:
      type: object
      properties:
        ticker:
          type: string
        action:
          type: string
          description: "buy, sell, or hold"
        price:
          type: number
        confidence:
          type: number
        target:
          type: number
        stop:
          type: number
        message:
          type: string
        timeframe:
          type: string
      required:
        - ticker
        - action
        - price