# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Dashboard API
 * OpenAPI spec version: 0.2.0
 */
import type { AssetAssetClass } from "./assetAssetClass";

export interface Asset {
  id: number;
  symbol: string;
  name: string;
  assetClass: AssetAssetClass;
  currentPrice: number;
  change: number;
  changePercent: number;
  volume: string;
  marketCap: string;
}