# lib/api-zod/src/generated/types/newPosition.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Dashboard API
 * OpenAPI spec version: 0.2.0
 */
import type { NewPositionAssetClass } from "./newPositionAssetClass";
import type { NewPositionDirection } from "./newPositionDirection";

export interface NewPosition {
  symbol: string;
  name: string;
  assetClass: NewPositionAssetClass;
  direction: NewPositionDirection;
  quantity: number;
  entryPrice: number;
  targetPrice: number;
  stopLoss: number;
}